requirements.txt
biopython==1.79
numpy==1.22.3
pysam==0.19.0
